fix: Fix "pairing" of notifications in Fabric
Previously, Fabric sent notifications in pairs if there was sufficient delay between the notifications, due to Ignite processing the local queries for the cache before we could write to the queue.
Note that this commit introduces a potential race condition if Fabric reads the queued notification, sends that to the runtime, and the runtime tries to read the notification, all before we can write it to the cache.