refactor!: Resolve AsyncLocals.Instance confusion
Initally AsyncLocals.Instance was meant to be equivalent to {Call,Stream}Data.Instance, that is, the agent instance.
However, when implementing Stream, it got reused for StreamTriggerNotification.Stream, that is, the stream "instance".
Later implemention of Perper.Model.Program and later PerperStartup followed that practice of setting instance to *TriggerNotification.{Call,Stream}.
However, agent instance happens to be important when tracking per-instance state. As we don't want StreamItemNotifcations to be queued across the whole agent instance, adding an extra property, AsyncLocals.Execution, was the best way forward.
Naming: StreamTriggerNotification and CallTrigerNotification use Stream and Call to name this field. As StreamOrCall was not a particularly inspiring name, and Instance was already taken (in {Call,Stream}Data), I decided to go with Execution.