fix: Workaround a weird bug in Ignite
(aka, if it works, don't touch it. BinaryObjectBuilder.GetField uses offsets from the start of the object for caching fields, while BinaryObjectBuilder.Mutate0 uses positions from the start of the stream. By going to builder and back to binary object, we create a binary object that starts at offset 0, which just happens to make the two types of positions equivalent.)