feat: Implement execution reservations
Note: The resulting execution distribution is not very fair on a single node (likely due to query result order being biased).
Also, ideally there would be a better way to do reservations than spawning a dedicated thread for each lock. (We can't easily reuse that, as we need the same thread for unlocking.)